# TAN-1813 - Permissions performance (2)