--- a/back/app/controllers/web_api/v1/followers_controller.rb
+++ b/back/app/controllers/web_api/v1/followers_controller.rb
@@ -2,9 +2,10 @@
 
 class WebApi::V1::FollowersController < ApplicationController
   def index
-    followers = policy_scope current_user.follows.includes(:followable)
+    followers = policy_scope current_user.follows
     followers = followers.where(followable_type: params[:followable_type]) if params.key? :followable_type
     followers = paginate followers.order(created_at: :desc)
+    followers = followers.includes(:followable)
     render json: linked_json(followers, WebApi::V1::FollowerSerializer, params: jsonapi_serializer_params, include: [:followable])
   end