--- a/back/app/controllers/web_api/v1/ideas_controller.rb
+++ b/back/app/controllers/web_api/v1/ideas_controller.rb
@@ -381,12 +381,12 @@ class WebApi::V1::IdeasController < ApplicationController
         end
       user_followers ||= {}
       {
-        params: jsonapi_serializer_params(vbii: reactions.index_by(&:reactable_id), user_followers: user_followers, pcs: Permissions::IdeaPermissionsService.new),
+        params: jsonapi_serializer_params(vbii: reactions.index_by(&:reactable_id), user_followers: user_followers),
         include: include
       }
     else
       {
-        params: jsonapi_serializer_params(pcs: Permissions::IdeaPermissionsService.new),
+        params: jsonapi_serializer_params,
         include: include
       }
     end