--- a/back/app/controllers/web_api/v1/ideas_controller.rb
+++ b/back/app/controllers/web_api/v1/ideas_controller.rb
@@ -21,18 +21,23 @@ class WebApi::V1::IdeasController < ApplicationController
     ideas = IdeasFinder.new(
       params,
       scope: policy_scope(Idea).where(publication_status: 'published'),
-      current_user: current_user,
-      includes: [
-        :idea_images, :idea_trending_info, :topics,
-        :idea_import, # defined through BulkImportIdeas engine
-        {
-          project: [:phases, { custom_form: [:custom_fields] }],
-          phases: [:permissions],
-          author: [:unread_notifications]
-        }
-      ]
+      current_user: current_user
     ).find_records
+
     ideas = paginate SortByParamsService.new.sort_ideas(ideas, params, current_user)
+
+    # Only include after pagination - so we only get associations we need
+    ideas = ideas.includes(
+      :idea_images,
+      :idea_trending_info,
+      :topics,
+      :idea_import, # defined through BulkImportIdeas engine
+      {
+        project: [:phases, { custom_form: [:custom_fields] }],
+        phases: [:permissions],
+        author: [:unread_notifications]
+      }
+    )
 
     ideas = convert_phase_voting_counts ideas, params
 
@@ -47,6 +52,7 @@ class WebApi::V1::IdeasController < ApplicationController
       includes: %i[idea_trending_info]
     ).find_records
     ideas = paginate SortByParamsService.new.sort_ideas(ideas, params, current_user)
+    ideas = ideas.includes(:idea_trending_info)
 
     render json: linked_json(ideas, WebApi::V1::IdeaMiniSerializer, params: jsonapi_serializer_params)
   end
@@ -55,10 +61,10 @@ class WebApi::V1::IdeasController < ApplicationController
     ideas = IdeasFinder.new(
       params,
       scope: policy_scope(Idea).where(publication_status: 'published'),
-      current_user: current_user,
-      includes: %i[author topics project idea_status idea_files]
+      current_user: current_user
     ).find_records
     ideas = paginate SortByParamsService.new.sort_ideas(ideas, params, current_user)
+    ideas = ideas.includes(:author, :topics, :project, :idea_status, :idea_files)
 
     render json: linked_json(ideas, WebApi::V1::PostMarkerSerializer, params: jsonapi_serializer_params)
   end
@@ -67,10 +73,10 @@ class WebApi::V1::IdeasController < ApplicationController
     ideas = IdeasFinder.new(
       params.merge(filter_can_moderate: true),
       scope: policy_scope(Idea).where(publication_status: 'published'),
-      current_user: current_user,
-      includes: %i[author topics project idea_status idea_files]
+      current_user: current_user
     ).find_records
     ideas = SortByParamsService.new.sort_ideas(ideas, params, current_user)
+    ideas = ideas.includes(:author, :topics, :project, :idea_status, :idea_files)
 
     I18n.with_locale(current_user&.locale) do
       xlsx = XlsxService.new.generate_ideas_xlsx ideas, view_private_attributes: true
@@ -82,10 +88,10 @@ class WebApi::V1::IdeasController < ApplicationController
     all_ideas = IdeasFinder.new(
       params,
       scope: policy_scope(Idea),
-      current_user: current_user,
-      includes: %i[idea_trending_info]
+      current_user: current_user
     ).find_records
     all_ideas = paginate SortByParamsService.new.sort_ideas(all_ideas, params, current_user)
+    all_ideas = all_ideas.includes(:idea_trending_info)
     counts = {
       'idea_status_id' => {},
       'topic_id' => {}