--- a/back/app/models/concerns/user_confirmation.rb
+++ b/back/app/models/concerns/user_confirmation.rb
@@ -90,6 +90,6 @@ module UserConfirmation
   end
 
   def user_confirmation_enabled?
-    AppConfiguration.instance.feature_activated?('user_confirmation')
+    @user_confirmation_enabled ||= AppConfiguration.instance.feature_activated?('user_confirmation')
   end
 end