--- a/back/app/policies/idea_comment_policy.rb
+++ b/back/app/policies/idea_comment_policy.rb
@@ -52,7 +52,6 @@ class IdeaCommentPolicy < ApplicationPolicy
   private
 
   def check_commenting_allowed(comment, user)
-    pcs = Permissions::IdeaPermissionsService.new
-    !pcs.denied_reason_for_action 'commenting_idea', user, comment.post
+    !Permissions::IdeaPermissionsService.new.denied_reason_for_action 'commenting_idea', user, comment.post
   end
 end