--- a/back/app/policies/idea_policy.rb
+++ b/back/app/policies/idea_policy.rb
@@ -67,8 +67,8 @@ class IdeaPolicy < ApplicationPolicy
     return true if (record.draft? && owner?) || (user && UserRoleService.new.can_moderate_project?(record.project, user))
     return false unless active? && owner? && ProjectPolicy.new(user, record.project).show?
 
-    pcs_posting_reason = Permissions::ProjectPermissionsService.new.denied_reason_for_action 'posting_idea', user, record.project
-    raise_not_authorized(pcs_posting_reason) if pcs_posting_reason && EXCLUDED_REASONS_FOR_UPDATE.exclude?(pcs_posting_reason)
+    posting_denied_reason = Permissions::ProjectPermissionsService.new.denied_reason_for_action 'posting_idea', user, record.project
+    raise_not_authorized(posting_denied_reason) if posting_denied_reason && EXCLUDED_REASONS_FOR_UPDATE.exclude?(posting_denied_reason)
     true
   end