--- a/back/app/serializers/web_api/v1/project_serializer.rb
+++ b/back/app/serializers/web_api/v1/project_serializer.rb
@@ -32,7 +32,7 @@ class WebApi::V1::ProjectSerializer < WebApi::V1::BaseSerializer
   end
 
   attribute :action_descriptor do |object, params|
-    @project_permissions_service ||= Permissions::ProjectPermissionsService.new
+    @project_permissions_service = params[:permissions_service] || Permissions::ProjectPermissionsService.new
     @project_permissions_service.action_descriptors object, current_user(params)
   end