--- a/back/engines/commercial/bulk_import_ideas/app/serializers/bulk_import_ideas/extensions/web_api/v1/idea_serializer.rb
+++ b/back/engines/commercial/bulk_import_ideas/app/serializers/bulk_import_ideas/extensions/web_api/v1/idea_serializer.rb
@@ -7,7 +7,9 @@ module BulkImportIdeas
         module IdeaSerializer
           def self.included(base)
             base.class_eval do
-              has_one :idea_import, serializer: BulkImportIdeas::WebApi::V1::IdeaImportSerializer
+              has_one :idea_import, serializer: BulkImportIdeas::WebApi::V1::IdeaImportSerializer, if: proc { |object, params|
+                can_moderate?(object, params)
+              }
             end
           end
         end