--- a/back/engines/commercial/public_api/app/controllers/public_api/v1/projects_controller.rb
+++ b/back/engines/commercial/public_api/app/controllers/public_api/v1/projects_controller.rb
@@ -7,11 +7,10 @@ module PublicApi
     def index
       @projects = PublicApi::ProjectPolicy::Scope.new(current_public_api_api_client, Project).resolve
       @projects = @projects
-        .includes(:project_images)
-        .includes(:map_config)
         .order(created_at: :desc)
         .page(params[:page_number])
         .per([params[:page_size]&.to_i || 12, 24].min)
+        .includes(:project_images, :map_config)
 
       render json: @projects,
         each_serializer: V1::ProjectSerializer,