--- a/back/engines/commercial/public_api/app/controllers/public_api/v2/baskets_controller.rb
+++ b/back/engines/commercial/public_api/app/controllers/public_api/v2/baskets_controller.rb
@@ -5,7 +5,7 @@ module PublicApi
     include DeletedItemsAction
 
     def index
-      list_items Basket.all.includes(:phase), V2::BasketSerializer
+      list_items Basket.all, V2::BasketSerializer, includes: [:phase]
     end
 
     def show