--- a/back/engines/commercial/verification/app/services/verification/patches/permissions/base_permissions_service.rb
+++ b/back/engines/commercial/verification/app/services/verification/patches/permissions/base_permissions_service.rb
@@ -4,6 +4,9 @@ module Verification
   module Patches
     module Permissions
       module BasePermissionsService
+
+        private
+
         def denied_when_permitted_by_groups?(permission, user)
           denied_reason = super