… back/engines/commercial/public_api/app/controllers/public_api/v2/comments_controller.rb
… back/engines/commercial/public_api/app/controllers/public_api/v2/ideas_controller.rb
… back/engines/commercial/public_api/app/controllers/public_api/v2/phases_controller.rb
… back/engines/commercial/public_api/app/controllers/public_api/v2/project_folders_controller.rb
… back/engines/commercial/public_api/app/controllers/public_api/v2/projects_controller.rb
… back/engines/commercial/public_api/app/controllers/public_api/v2/reactions_controller.rb
… back/engines/commercial/public_api/app/controllers/public_api/v2/volunteering_causes_controller.rb
… back/engines/free/polls/app/controllers/polls/web_api/v1/questions_controller.rb
… back/engines/free/polls/app/policies/polls/response_policy.rb
… back/engines/free/volunteering/app/controllers/volunteering/web_api/v1/volunteers_controller.rb
… back/spec/acceptance/projects_spec.rb
… back/spec/services/permissions/idea_permissions_service_spec.rb
… back/spec/services/permissions/project_permissions_service_spec.rb
… front/cypress/e2e/project_page_poll.cy.ts