# Second set of performance improvements
Further reduced the number of queries - particularly for users/group permissions
- Queries reduced from an initial 500ish (in the refactor PR) to single figures for the tests in this PR
- Memoized app configuration in User model
- Memoized groups and smart groups because of the number of times groups potentially get checked in action descriptors
- Passed project down to the base service so that it could make use of the includes array when checking user moderator status 
- Reconfigured where permissions were included in the includes array for the ideas controller 

Went a little further than just the permissions as I spotted a performance hit with paging in the ideas controller and then realised there were many other places this could be improved:
- Ensured that `includes` is always called after pagination in controllers - this avoids rails preloading items it does not need

Also, ran some performance tests locally in postman against this branch and also against master (so it tests the performance improvements in this branch and the previous one `TAN-1813-permissions-performance`). The endpoints were called in the tests with paging as they are within the application.

Definitely improves performance - but by differing amounts depending on the data in the tenant. Test results based on the average of 10 requests logged in as admin:

### Localhost - which I'd been using to test idea imports - 10 projects + 250 ideas
- `/web_api/v1/ideas` - master: 59385ms - this branch: 827ms - 98% decrease
- `/web_api/v1/projects` - master 1263ms - this branch: 866ms - 31% decrease

### strategicplan.citizenlab.co - imported from production - 37 projects + 165 ideas
- `/web_api/v1/ideas` - master: 1442ms - this branch: 665ms - 53% decrease
- `/web_api/v1/projects` - master: 2544ms - this branch: 2168ms - 15% decrease

# Changelog
## Technical
- Performance improvements to action descriptors and therefore projects and ideas endpoints